Remove some debug code. Add some api call
