+++ client/src/lib/api.ts
@@ -16,7 +16,6 @@ const axiosInstance = axios.create({
 axiosInstance.interceptors.request.use(
   (config: InternalAxiosRequestConfig<any>) => {
     const token = localStorage.getItem("token");
-    // console.log("token:", token);
     if (token) {
       config.headers["Authorization"] = `Bearer ${token}`;
     }
