+++ client/src/pages/poster/index.tsx
@@ -1,23 +1,92 @@
 import jwt from "jsonwebtoken";
+import Image from "next/image";
+import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 
-import { PosterTaskCardProps } from "@/components/ui/poster/task-card";
+import { Button } from "@/components/ui/button";
+import { ErrorCallout } from "@/components/ui/callout";
+import PosterTaskCard, {
+  PosterTaskCardProps,
+} from "@/components/ui/poster/task-card";
+import TaskTopBar from "@/components/ui/task-top-bar";
 import { axiosInstance } from "@/lib/api";
 
-import Create from "./tasks/create-task";
-import TaskList from "./tasks/task-list";
+export function Create() {
+  const router = useRouter();
+  return (
+    <div className="flex flex-col items-center">
+      <TaskTopBar />
+      <Image
+        className="mt-52"
+        src="/task-blank-page.svg"
+        alt="No task"
+        width={200}
+        height={200}
+      />
+      <p className="mt-10 font-bold">You do not have any orders yet</p>
+      <Button
+        className="mt-10 w-11/12"
+        label="Create a Task"
+        onClick={() => router.push("/poster/tasks/new-task")}
+      ></Button>
+    </div>
+  );
+}
+
+type TaskListProp = {
+  className?: string;
+  tasks: any[];
+};
+
+function TaskList({ className, tasks }: TaskListProp) {
+  const router = useRouter();
+  // TODO api fetch tasklist replace dummy data.in reverse order.
+  // pass taskid and title to next page.
+
+  return (
+    <div
+      className={`m-4 flex flex-col items-center justify-between ${className}`}
+    >
+      <TaskTopBar />
+      <div className="w-full">
+        {tasks.map((task, index) => (
+          <div className={`mt-1`} key={index}>
+            <PosterTaskCard
+              title={task.title}
+              numberOfBidders={5} //need to get bidder data
+              date={task.deadline.slice(0, 10)}
+              state={task.status}
+              onClick={() => {
+                router.push({
+                  pathname: `/poster/tasks/${task.task_id}/task-details`,
+                });
+              }}
+            ></PosterTaskCard>
+          </div>
+        ))}
+      </div>
+      <Button
+        size="floating"
+        variant="floating"
+        onClick={() => router.push("/poster/tasks/new-task")}
+        className={`fixed bottom-10 right-10`}
+      >
+        + New Task
+      </Button>
+    </div>
+  );
+}
 
 export default function PosterTasksPage() {
   const [tasks, setTasks] = useState<object[]>([]);
   const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+  const [error, setError] = useState<string>("");
 
   useEffect(() => {
     const fetchUserData = async () => {
       try {
         const response = await axiosInstance.get("/app/tasks/");
         const jsonResponse = response.data;
-        // console.log(jsonResponse);
 
         const token = localStorage.getItem("token");
         if (!token) throw new Error("No token found");
@@ -32,10 +101,8 @@ export default function PosterTasksPage() {
         setTasks(filteredTasks);
       } catch (error: unknown) {
         if (error instanceof Error) {
-          console.error("Error fetching user data:", error.message);
           setError(error.message);
         } else {
-          console.error("An unexpected error occurred:", error);
           setError("An unexpected error occurred");
         }
       } finally {
@@ -50,5 +117,15 @@ export default function PosterTasksPage() {
 
   const hasTask = tasks.length > 0;
 
-  return <div>{!hasTask ? <Create /> : <TaskList />}</div>;
+  return (
+    <div>
+      {error != "" ? (
+        <ErrorCallout className="pb-4" text={error} />
+      ) : !hasTask ? (
+        <Create />
+      ) : (
+        <TaskList tasks={tasks} />
+      )}
+    </div>
+  );
 }
